% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data.R
\docType{data}
\name{amino_acids}
\alias{amino_acids}
\title{List of Amino Acids}
\format{
A character vector with 20 elements.
}
\usage{
amino_acids
}
\description{
This variable contains a list of amino acids used for validation in the \code{tRNA_Gene} class.
}
\examples{
tRNA1 <- tRNA_Gene(id = 1, specific_aa = "Alanine")

}
\keyword{datasets}
