% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_Gene.R
\name{category<-,Gene-method}
\alias{category<-,Gene-method}
\title{Set category attribute}
\usage{
\S4method{category}{Gene}(x) <- value
}
\arguments{
\item{x}{An object of class Gene.}

\item{value}{The value to set for the category attribute.}
}
\value{
The modified object.
}
\description{
Setter for category attribute of Gene class.
}
