% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_Gene.R
\name{category}
\alias{category}
\title{Get category attribute}
\usage{
category(x)
}
\arguments{
\item{x}{An object of class Gene.}
}
\value{
The value of the category attribute.
}
\description{
Getter for category attribute of Gene class.
}
