% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_CodingGene.R
\name{CDS_end,CodingGene-method}
\alias{CDS_end,CodingGene-method}
\title{Get CDS_end attribute}
\usage{
\S4method{CDS_end}{CodingGene}(x)
}
\arguments{
\item{x}{An object of class CodingGene.}
}
\value{
The value of the CDS_end attribute.
}
\description{
Getter for CDS_end attribute of CodingGene class.
}
