% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_CodingGene.R
\name{CDS_end<-,CodingGene-method}
\alias{CDS_end<-,CodingGene-method}
\title{Set CDS_end attribute}
\usage{
\S4method{CDS_end}{CodingGene}(x) <- value
}
\arguments{
\item{x}{An object of class CodingGene.}

\item{value}{The value to set for the CDS_end attribute.}
}
\value{
The modified object.
}
\description{
Setter for CDS_end attribute of CodingGene class.
}
