% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_CodingGene.R
\name{CDS_end}
\alias{CDS_end}
\title{Get CDS_end attribute}
\usage{
CDS_end(x)
}
\arguments{
\item{x}{An object of class CodingGene.}
}
\value{
The value of the CDS_end attribute.
}
\description{
Getter for CDS_end attribute of CodingGene class.
}
