% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_CodingGene.R
\name{CDS_start<-}
\alias{CDS_start<-}
\title{Set CDS_start attribute}
\usage{
CDS_start(x) <- value
}
\arguments{
\item{x}{An object of class CodingGene.}

\item{value}{The value to set for the CDS_start attribute.}
}
\value{
The modified object.
}
\description{
Setter for CDS_start attribute of CodingGene class.
}
