% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_CodingGene.R
\name{CDS_start}
\alias{CDS_start}
\title{Get CDS_start attribute}
\usage{
CDS_start(x)
}
\arguments{
\item{x}{An object of class CodingGene.}
}
\value{
The value of the CDS_start attribute.
}
\description{
Getter for CDS_start attribute of CodingGene class.
}
