% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_Gene.R
\name{chrom,Gene-method}
\alias{chrom,Gene-method}
\title{Get chrom attribute}
\usage{
\S4method{chrom}{Gene}(x)
}
\arguments{
\item{x}{An object of class Gene.}
}
\value{
The value of the chrom attribute.
}
\description{
Getter for chrom attribute of Gene class.
}
