% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_Gene.R
\name{chrom<-}
\alias{chrom<-}
\title{Set chrom attribute}
\usage{
chrom(x) <- value
}
\arguments{
\item{x}{An object of class Gene.}

\item{value}{The value to set for the chrom attribute.}
}
\value{
The modified object.
}
\description{
Setter for chrom attribute of Gene class.
}
