% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{chrom}
\alias{chrom}
\title{Get chrom attribute}
\usage{
chrom(x)
}
\arguments{
\item{x}{An object of class Gene.}
}
\value{
The value of the chrom attribute.
}
\description{
Getter for chrom attribute of Gene class.
}
