% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_data2.R
\docType{data}
\name{classList}
\alias{classList}
\title{List of all classes}
\format{
A character list with 10 elements.
}
\usage{
classList
}
\description{
This variable contains the list of all classes to be used in the
\code{createAccessors} function.
}
\examples{
print(classList)

}
\keyword{datasets}
