% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_data2.R
\docType{data}
\name{classList_for_product}
\alias{classList_for_product}
\title{List of all classes except the virtual one}
\format{
A character list with 10 elements.
}
\usage{
classList_for_product
}
\description{
This variable contains the list of all classes except the virtual one to be used in the \code{LengthProduct} function.
}
\examples{
print(classList_for_product)
}
\keyword{datasets}
