% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_coding_gene_class.R
\docType{class}
\name{CodingGene-class}
\alias{CodingGene-class}
\alias{CodingGene}
\title{CodingGene Class}
\value{
A \code{CodingGene} object.
}
\description{
A class to represent a protein-coding gene.
}
\details{
Additional details:
The prototype are:
CDS_start = NA_real_, CDS_end = NA_real_, length_protein = NA_real_,
protein_ID = NA_real_, gene_product = "Protein",
category = "Protein-coding".

The validity function checks the validity of the object by ensuring:
\itemize{
\item The length of the protein must be at least 100 amino acids.
\item If the gene is annotated on the positive strand, CDS_start must be
lower than CDS_end.
\item If the gene is annotated on the negative strand, CDS_start must be
higher than CDS_end.
\item If the gene is annotated on the positive strand, CDS_start must be
higher than the TSS.
\item If the gene is annotated on the negative strand, CDS_start must be
lower than the TSS.
\item If the gene is annotated on the negative strand, CDS_end must be
higher than the TTS.
\item If the gene is annotated on the positive strand, CDS_end must be
lower than the TTS.
}
}
\section{Slots}{

\describe{
\item{\code{CDS_start}}{A numeric value representing the coding sequence start
position.}

\item{\code{CDS_end}}{A numeric value representing the coding sequence end position.}

\item{\code{protein_ID}}{A numeric value representing the protein ID.}

\item{\code{length_protein}}{A numeric value representing the length of the protein
in amino acids.}

\item{\code{category}}{A character string representing the category of the gene.
Default is "Protein-coding" and it's not possibile to change it.}

\item{\code{gene_product}}{A character string representing the gene product.
Default is "Protein" and it's not possibile to change it.}
}}

\section{Author}{

Elena Sasso \email{elena.sasso@mail.polimi.it}
}

\examples{
# Create a CodingGene object
Gene1 <- CodingGene(id = 1, Hugo_symbol = "TP53", chrom = "17", 
length_protein = 393)

}
\seealso{
\link[=Gene-class]{Gene}
}
