% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_Functions.R
\name{ComputeGeneLength,Gene-method}
\alias{ComputeGeneLength,Gene-method}
\title{ComputeGeneLength method for Gene class}
\usage{
\S4method{ComputeGeneLength}{Gene}(.object)
}
\arguments{
\item{.object}{An object of class \code{Gene}.}
}
\value{
A numeric value representing the length of the gene.
}
\description{
This method calculates the length of a gene in base pairs for objects of class \code{Gene}.
}
\examples{
Brca1 <- CodingGene(id = 1, Hugo_symbol = "BRCA1", chrom = "17", start = 41196312, end = 41277500)
ComputeGeneLength(Brca1)
}
