% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_Functions.R
\name{ComputeGeneLength}
\alias{ComputeGeneLength}
\title{Compute Gene Length}
\usage{
ComputeGeneLength(.object)
}
\arguments{
\item{.object}{An object of class \code{Gene}.}
}
\value{
A numeric value representing the length of the gene.
}
\description{
This function calculates the length of a gene in base pairs based on its
start and end positions.
}
