% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_Gene.R
\name{description,Gene-method}
\alias{description,Gene-method}
\title{Get description attribute}
\usage{
\S4method{description}{Gene}(x)
}
\arguments{
\item{x}{An object of class Gene.}
}
\value{
The value of the description attribute.
}
\description{
Getter for description attribute of Gene class.
}
