% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_Gene.R
\name{description<-,Gene-method}
\alias{description<-,Gene-method}
\title{Set description attribute}
\usage{
\S4method{description}{Gene}(x) <- value
}
\arguments{
\item{x}{An object of class Gene.}

\item{value}{The value to set for the description attribute.}
}
\value{
The modified object.
}
\description{
Setter for description attribute of Gene class.
}
