% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{description}
\alias{description}
\title{Get description attribute}
\usage{
description(x)
}
\arguments{
\item{x}{An object of class Gene.}
}
\value{
The value of the description attribute.
}
\description{
Getter for description attribute of Gene class.
}
