% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_Gene.R
\name{end<-,Gene-method}
\alias{end<-,Gene-method}
\title{Set end attribute}
\usage{
\S4method{end}{Gene}(x) <- value
}
\arguments{
\item{x}{An object of class Gene.}

\item{value}{The value to set for the end attribute.}
}
\value{
The modified object.
}
\description{
Setter for end attribute of Gene class.
}
