% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{end<-}
\alias{end<-}
\title{Set end attribute}
\usage{
end(x) <- value
}
\arguments{
\item{x}{An object of class Gene.}

\item{value}{The value to set for the end attribute.}
}
\value{
The modified object.
}
\description{
Setter for end attribute of Gene class.
}
