% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{end}
\alias{end}
\title{Get end attribute}
\usage{
end(x, ...)
}
\arguments{
\item{x}{An object of class Gene.}
}
\value{
The value of the end attribute.
}
\description{
Getter for end attribute of Gene class.
}
