% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_Gene.R
\name{gene_product,Gene-method}
\alias{gene_product,Gene-method}
\title{Get gene_product attribute}
\usage{
\S4method{gene_product}{Gene}(x)
}
\arguments{
\item{x}{An object of class Gene.}
}
\value{
The value of the gene_product attribute.
}
\description{
Getter for gene_product attribute of Gene class.
}
