% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_Gene.R
\name{gene_product<-,Gene-method}
\alias{gene_product<-,Gene-method}
\title{Set gene_product attribute}
\usage{
\S4method{gene_product}{Gene}(x) <- value
}
\arguments{
\item{x}{An object of class Gene.}

\item{value}{The value to set for the gene_product attribute.}
}
\value{
The modified object.
}
\description{
Setter for gene_product attribute of Gene class.
}
