% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{gene_product<-}
\alias{gene_product<-}
\title{Set gene_product attribute}
\usage{
gene_product(x) <- value
}
\arguments{
\item{x}{An object of class Gene.}

\item{value}{The value to set for the gene_product attribute.}
}
\value{
The modified object.
}
\description{
Setter for gene_product attribute of Gene class.
}
