% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{gene_product}
\alias{gene_product}
\title{Get gene_product attribute}
\usage{
gene_product(x)
}
\arguments{
\item{x}{An object of class Gene.}
}
\value{
The value of the gene_product attribute.
}
\description{
Getter for gene_product attribute of Gene class.
}
