% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_getter_setter.R
\name{generateGetterSetterCode_for_Gene}
\alias{generateGetterSetterCode_for_Gene}
\title{Generate Getter and Setter Code for Gene class}
\usage{
generateGetterSetterCode_for_Gene(class_name, file_path)
}
\arguments{
\item{class_name}{The name of the S4 class for which the getter and setter
methods are to be generated.}

\item{file_path}{The file path where the generated code will be written.}
}
\value{
This function does not return a value. It writes the generated
code directly to the specified file.
}
\description{
Generate Getter and Setter Code for Gene class
}
\details{
This function generates Roxygen2 documentation for each getter
and setter method, making them ready for inclusion in an R package.
It writes the documentations and the methods in the appropriate files,
which were empty at the beginning.
For attributes named "gene_product" and "category",
the setter methods are generated with a restriction that prevents
modification of these attributes, providing an error message instead.
}
