% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_getter_setter.R
\name{generateGetterSetterCode_for_others}
\alias{generateGetterSetterCode_for_others}
\title{Generate Getter and Setter Code for derived classes
This function generates getter and setter methods for all attributes of the
derived class, writing the code to a specified file. It includes additional
restrictions for certain attributes where the setter method is not allowed
to modify the attribute.}
\usage{
generateGetterSetterCode_for_others(class_name, file_path)
}
\arguments{
\item{class_name}{The name of the S4 class for which the getter and setter
methods are to be generated.}

\item{file_path}{The file path where the generated code will be written.}
}
\value{
This function does not return a value. It writes the generated
code directly to the specified file.
}
\description{
Generate Getter and Setter Code for derived classes
This function generates getter and setter methods for all attributes of the
derived class, writing the code to a specified file. It includes additional
restrictions for certain attributes where the setter method is not allowed
to modify the attribute.
}
\details{
This function generates Roxygen2 documentation for each getter
and setter method, making them ready for inclusion in an R package.
For attributes named "type_RNA" and "types_RNA",
the setter methods are generated with a restriction that prevents
modification of these attributes, providing an error message instead.
}
