% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Housekeeping_tRNA_rRNA_classes.R
\docType{class}
\name{Housekeeping_RNA_Gene-class}
\alias{Housekeeping_RNA_Gene-class}
\alias{Housekeeping_RNA_Gene}
\title{Housekeeping_RNA Genes Class}
\value{
A \code{Housekeeping_RNA_Gene} object.
}
\description{
A class to represent housekeeping genes.
}
\details{
Additional details:
The prototype for the category is "Housekeeping".

The validity function checks the validity of the object by ensuring that the
category is right
}
\section{Slots}{

\describe{
\item{\code{category}}{A character string representing the category of the gene.
Default is "Housekeeping" and it's not possibile to change it.}
}}

\section{Author}{

Elena Sasso \email{elena.sasso@mail.polimi.it}
}

\examples{
# Create a housekeeping_RNA_Gene object
Housekeeping1 <- Housekeeping_RNA_Gene(id = 2)

}
\seealso{
\link[=Gene-class]{Gene}
}
