% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_Gene.R
\name{Hugo_symbol,Gene-method}
\alias{Hugo_symbol,Gene-method}
\title{Get Hugo_symbol attribute}
\usage{
\S4method{Hugo_symbol}{Gene}(x)
}
\arguments{
\item{x}{An object of class Gene.}
}
\value{
The value of the Hugo_symbol attribute.
}
\description{
Getter for Hugo_symbol attribute of Gene class.
}
