% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_Gene.R
\name{Hugo_symbol<-,Gene-method}
\alias{Hugo_symbol<-,Gene-method}
\title{Set Hugo_symbol attribute}
\usage{
\S4method{Hugo_symbol}{Gene}(x) <- value
}
\arguments{
\item{x}{An object of class Gene.}

\item{value}{The value to set for the Hugo_symbol attribute.}
}
\value{
The modified object.
}
\description{
Setter for Hugo_symbol attribute of Gene class.
}
