% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_Gene.R
\name{Hugo_symbol<-}
\alias{Hugo_symbol<-}
\title{Set Hugo_symbol attribute}
\usage{
Hugo_symbol(x) <- value
}
\arguments{
\item{x}{An object of class Gene.}

\item{value}{The value to set for the Hugo_symbol attribute.}
}
\value{
The modified object.
}
\description{
Setter for Hugo_symbol attribute of Gene class.
}
