% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_Gene.R
\name{id,Gene-method}
\alias{id,Gene-method}
\title{Get id attribute}
\usage{
\S4method{id}{Gene}(x)
}
\arguments{
\item{x}{An object of class Gene.}
}
\value{
The value of the id attribute.
}
\description{
Getter for id attribute of Gene class.
}
