% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{id<-}
\alias{id<-}
\title{Set id attribute}
\usage{
id(x) <- value
}
\arguments{
\item{x}{An object of class Gene.}

\item{value}{The value to set for the id attribute.}
}
\value{
The modified object.
}
\description{
Setter for id attribute of Gene class.
}
