% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{id}
\alias{id}
\title{Get id attribute}
\usage{
id(x)
}
\arguments{
\item{x}{An object of class Gene.}
}
\value{
The value of the id attribute.
}
\description{
Getter for id attribute of Gene class.
}
