% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_lncRNA.R
\name{length_lncRNA<-}
\alias{length_lncRNA<-}
\title{Set length_lncRNA attribute}
\usage{
length_lncRNA(x) <- value
}
\arguments{
\item{x}{An object of class lncRNA_Gene.}

\item{value}{The value to set for the length_lncRNA attribute.}
}
\value{
The modified object.
}
\description{
Setter for length_lncRNA attribute of lncRNA_Gene class.
}
