% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_lncRNA.R
\name{length_lncRNA}
\alias{length_lncRNA}
\title{Get length_lncRNA attribute}
\usage{
length_lncRNA(x)
}
\arguments{
\item{x}{An object of class lncRNA_Gene.}
}
\value{
The value of the length_lncRNA attribute.
}
\description{
Getter for length_lncRNA attribute of lncRNA_Gene class.
}
