% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_CodingGene.R
\name{length_protein<-,CodingGene-method}
\alias{length_protein<-,CodingGene-method}
\title{Set length_protein attribute}
\usage{
\S4method{length_protein}{CodingGene}(x) <- value
}
\arguments{
\item{x}{An object of class CodingGene.}

\item{value}{The value to set for the length_protein attribute.}
}
\value{
The modified object.
}
\description{
Setter for length_protein attribute of CodingGene class.
}
