% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_CodingGene.R
\name{length_protein}
\alias{length_protein}
\title{Get length_protein attribute}
\usage{
length_protein(x)
}
\arguments{
\item{x}{An object of class CodingGene.}
}
\value{
The value of the length_protein attribute.
}
\description{
Getter for length_protein attribute of CodingGene class.
}
