% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_rRNA.R
\name{length_rRNA,rRNA_Gene-method}
\alias{length_rRNA,rRNA_Gene-method}
\title{Get length_rRNA attribute}
\usage{
\S4method{length_rRNA}{rRNA_Gene}(x)
}
\arguments{
\item{x}{An object of class rRNA_Gene.}
}
\value{
The value of the length_rRNA attribute.
}
\description{
Getter for length_rRNA attribute of rRNA_Gene class.
}
