% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_rRNA.R
\name{length_rRNA<-,rRNA_Gene-method}
\alias{length_rRNA<-,rRNA_Gene-method}
\title{Set length_rRNA attribute}
\usage{
\S4method{length_rRNA}{rRNA_Gene}(x) <- value
}
\arguments{
\item{x}{An object of class rRNA_Gene.}

\item{value}{The value to set for the length_rRNA attribute.}
}
\value{
The modified object.
}
\description{
Setter for length_rRNA attribute of rRNA_Gene class.
}
