% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_rRNA.R
\name{length_rRNA}
\alias{length_rRNA}
\title{Get length_rRNA attribute}
\usage{
length_rRNA(x)
}
\arguments{
\item{x}{An object of class rRNA_Gene.}
}
\value{
The value of the length_rRNA attribute.
}
\description{
Getter for length_rRNA attribute of rRNA_Gene class.
}
