% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_sncRNA.R
\name{length_sncRNA<-}
\alias{length_sncRNA<-}
\title{Set length_sncRNA attribute}
\usage{
length_sncRNA(x) <- value
}
\arguments{
\item{x}{An object of class sncRNA_Gene.}

\item{value}{The value to set for the length_sncRNA attribute.}
}
\value{
The modified object.
}
\description{
Setter for length_sncRNA attribute of sncRNA_Gene class.
}
