% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_sncRNA.R
\name{length_sncRNA}
\alias{length_sncRNA}
\title{Get length_sncRNA attribute}
\usage{
length_sncRNA(x)
}
\arguments{
\item{x}{An object of class sncRNA_Gene.}
}
\value{
The value of the length_sncRNA attribute.
}
\description{
Getter for length_sncRNA attribute of sncRNA_Gene class.
}
