% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_tRNA.R
\name{length_tRNA<-}
\alias{length_tRNA<-}
\title{Set length_tRNA attribute}
\usage{
length_tRNA(x) <- value
}
\arguments{
\item{x}{An object of class tRNA_Gene.}

\item{value}{The value to set for the length_tRNA attribute.}
}
\value{
The modified object.
}
\description{
Setter for length_tRNA attribute of tRNA_Gene class.
}
