% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_tRNA.R
\name{length_tRNA,tRNA_Gene-method}
\alias{length_tRNA,tRNA_Gene-method}
\title{Get length_tRNA attribute}
\usage{
\S4method{length_tRNA}{tRNA_Gene}(x)
}
\arguments{
\item{x}{An object of class tRNA_Gene.}
}
\value{
The value of the length_tRNA attribute.
}
\description{
Getter for length_tRNA attribute of tRNA_Gene class.
}
