% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_Functions.R
\name{lengthProductMethod}
\alias{lengthProductMethod}
\title{lengthProduct Methods for each class}
\usage{
lengthProductMethod(classes)
}
\arguments{
\item{classes}{A list of class names for which the length product methods
should be created.}
}
\value{
This function does not return a value. It creates methods for the
\code{lengthProduct} generic function.
}
\description{
This function creates specific methods for the \code{lengthProduct} generic
function for all classes.
}
\section{Author}{

Elena Sasso \email{elena.sasso@mail.polimi.it}
}

\examples{
Brca1 <- CodingGene(id = 1, Hugo_symbol = "BRCA1", chrom = "17", 
start = 41196312, end = 41277500)
lengthProduct(Brca1)

}
