% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_NON_Coding_gene_classes.R
\docType{class}
\name{lncRNA_Gene-class}
\alias{lncRNA_Gene-class}
\alias{lncRNA_Gene}
\title{lncRNA_Gene Class}
\value{
A \code{lncRNA_Gene} object.
}
\description{
A class to represent long non-coding RNA (lncRNA) genes.
}
\details{
Additional details:
The prototype are:
length_lncRNA = NA_real_, lncRNA_ID = NA_real_, gene_product = "lncRNA", type_RNA = "long".

The validity function checks the validity of the object by ensuring:
\itemize{
\item the length of the lncRNA is at least 200.
\item the product and the type(long) are correct.
}
}
\section{Slots}{

\describe{
\item{\code{length_lncRNA}}{A numeric value indicating the length of the lncRNA.}

\item{\code{lncRNA_ID}}{A numeric value representing the lncRNA ID.}

\item{\code{gene_product}}{A character string representing the product of the gene. Default is "lncRNA" and it's not possibile to change it.}

\item{\code{type_RNA}}{A character string representing the type of RNA. Default is "long" and it's not possible to change it.}
}}

\section{Author}{

Elena Sasso \email{elena.sasso@mail.polimi.it}
}

\examples{
# Create a lncRNA_Gene object
lncRNA_gene <- lncRNA_Gene(id = 6, Hugo_symbol = "lncFOXP3", length_lncRNA = 300, lncRNA_ID = 1)

}
\seealso{
\code{\link{ncRNA_Gene}}
}
