% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_lncRNA.R
\name{lncRNA_ID,lncRNA_Gene-method}
\alias{lncRNA_ID,lncRNA_Gene-method}
\title{Get lncRNA_ID attribute}
\usage{
\S4method{lncRNA_ID}{lncRNA_Gene}(x)
}
\arguments{
\item{x}{An object of class lncRNA_Gene.}
}
\value{
The value of the lncRNA_ID attribute.
}
\description{
Getter for lncRNA_ID attribute of lncRNA_Gene class.
}
