% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_lncRNA.R
\name{lncRNA_ID<-}
\alias{lncRNA_ID<-}
\title{Set lncRNA_ID attribute}
\usage{
lncRNA_ID(x) <- value
}
\arguments{
\item{x}{An object of class lncRNA_Gene.}

\item{value}{The value to set for the lncRNA_ID attribute.}
}
\value{
The modified object.
}
\description{
Setter for lncRNA_ID attribute of lncRNA_Gene class.
}
