% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_Gene.R
\name{location,Gene-method}
\alias{location,Gene-method}
\title{Get location attribute}
\usage{
\S4method{location}{Gene}(x)
}
\arguments{
\item{x}{An object of class Gene.}
}
\value{
The value of the location attribute.
}
\description{
Getter for location attribute of Gene class.
}
