% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_Gene.R
\name{location<-,Gene-method}
\alias{location<-,Gene-method}
\title{Set location attribute}
\usage{
\S4method{location}{Gene}(x) <- value
}
\arguments{
\item{x}{An object of class Gene.}

\item{value}{The value to set for the location attribute.}
}
\value{
The modified object.
}
\description{
Setter for location attribute of Gene class.
}
