% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{location}
\alias{location}
\title{Get location attribute}
\usage{
location(x)
}
\arguments{
\item{x}{An object of class Gene.}
}
\value{
The value of the location attribute.
}
\description{
Getter for location attribute of Gene class.
}
