% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_NON_Coding_gene_classes.R
\docType{class}
\name{miRNA_Gene-class}
\alias{miRNA_Gene-class}
\alias{miRNA_Gene}
\title{miRNA_Gene Class}
\value{
A \code{miRNA_Gene} object.
}
\description{
A class to represent microRNA (miRNA) genes.
}
\details{
Additional details:
The prototype are:
miRNA_ID = NA_real_, RNA_target = list(), gene_product = "miRNA".

The validity function ensures the gene product is 'miRNA'.
}
\section{Slots}{

\describe{
\item{\code{miRNA_ID}}{A numeric value representing the miRNA ID.}

\item{\code{RNA_target}}{A list of miRNA targets.}

\item{\code{gene_product}}{A character string representing the product of the gene.
Default is "miRNA" and it's not possibile to change it.}
}}

\section{Author}{

Elena Sasso \email{elena.sasso@mail.polimi.it}
}

\examples{
# Create a miRNA_Gene object
miRNA_gene <- miRNA_Gene(id = 8, miRNA_ID = 1, RNA_target = list("target1",
"target2"))

}
\seealso{
\code{\link{sncRNA_Gene}}
}
