% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_miRNA_Gene.R
\name{miRNA_ID<-}
\alias{miRNA_ID<-}
\title{Set miRNA_ID attribute}
\usage{
miRNA_ID(x) <- value
}
\arguments{
\item{x}{An object of class miRNA_Gene.}

\item{value}{The value to set for the miRNA_ID attribute.}
}
\value{
The modified object.
}
\description{
Setter for miRNA_ID attribute of miRNA_Gene class.
}
