% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_miRNA_Gene.R
\name{miRNA_ID}
\alias{miRNA_ID}
\title{Get miRNA_ID attribute}
\usage{
miRNA_ID(x)
}
\arguments{
\item{x}{An object of class miRNA_Gene.}
}
\value{
The value of the miRNA_ID attribute.
}
\description{
Getter for miRNA_ID attribute of miRNA_Gene class.
}
