% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_NON_Coding_gene_classes.R
\docType{class}
\name{ncRNA_Gene-class}
\alias{ncRNA_Gene-class}
\alias{ncRNA_Gene}
\title{ncRNA_Gene Class}
\value{
A \code{ncRNA_Gene} object.
}
\description{
A class to represent non-coding RNA (ncRNA) genes.
}
\details{
Additional details:
The prototype for category is "Regolative".

The validity function checks the validity of the object by ensuring that
the category is right.
}
\section{Slots}{

\describe{
\item{\code{category}}{A character string representing the category of the gene.
Default is "Regolative" and it's not possibile to change it.}
}}

\section{Author}{

Elena Sasso \email{elena.sasso@mail.polimi.it}
}

\examples{
# Create a ncRNA_Gene object
ncRNA_gene <- ncRNA_Gene(id = 5)

}
\seealso{
\link[=Gene-class]{Gene}
}
