% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_Gene.R
\name{number_exon<-,Gene-method}
\alias{number_exon<-,Gene-method}
\title{Set number_exon attribute}
\usage{
\S4method{number_exon}{Gene}(x) <- value
}
\arguments{
\item{x}{An object of class Gene.}

\item{value}{The value to set for the number_exon attribute.}
}
\value{
The modified object.
}
\description{
Setter for number_exon attribute of Gene class.
}
