% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_Gene.R
\name{number_exon<-}
\alias{number_exon<-}
\title{Set number_exon attribute}
\usage{
number_exon(x) <- value
}
\arguments{
\item{x}{An object of class Gene.}

\item{value}{The value to set for the number_exon attribute.}
}
\value{
The modified object.
}
\description{
Setter for number_exon attribute of Gene class.
}
