% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_Gene.R
\name{number_exon}
\alias{number_exon}
\title{Get number_exon attribute}
\usage{
number_exon(x)
}
\arguments{
\item{x}{An object of class Gene.}
}
\value{
The value of the number_exon attribute.
}
\description{
Getter for number_exon attribute of Gene class.
}
