% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_CodingGene.R
\name{protein_ID<-,CodingGene-method}
\alias{protein_ID<-,CodingGene-method}
\title{Set protein_ID attribute}
\usage{
\S4method{protein_ID}{CodingGene}(x) <- value
}
\arguments{
\item{x}{An object of class CodingGene.}

\item{value}{The value to set for the protein_ID attribute.}
}
\value{
The modified object.
}
\description{
Setter for protein_ID attribute of CodingGene class.
}
