% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_CodingGene.R
\name{protein_ID<-}
\alias{protein_ID<-}
\title{Set protein_ID attribute}
\usage{
protein_ID(x) <- value
}
\arguments{
\item{x}{An object of class CodingGene.}

\item{value}{The value to set for the protein_ID attribute.}
}
\value{
The modified object.
}
\description{
Setter for protein_ID attribute of CodingGene class.
}
