% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_CodingGene.R
\name{protein_ID}
\alias{protein_ID}
\title{Get protein_ID attribute}
\usage{
protein_ID(x)
}
\arguments{
\item{x}{An object of class CodingGene.}
}
\value{
The value of the protein_ID attribute.
}
\description{
Getter for protein_ID attribute of CodingGene class.
}
