% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_miRNA_Gene.R
\name{RNA_target<-}
\alias{RNA_target<-}
\title{Set RNA_target attribute}
\usage{
RNA_target(x) <- value
}
\arguments{
\item{x}{An object of class miRNA_Gene.}

\item{value}{The value to set for the RNA_target attribute.}
}
\value{
The modified object.
}
\description{
Setter for RNA_target attribute of miRNA_Gene class.
}
