% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_miRNA_Gene.R
\name{RNA_target}
\alias{RNA_target}
\title{Get RNA_target attribute}
\usage{
RNA_target(x)
}
\arguments{
\item{x}{An object of class miRNA_Gene.}
}
\value{
The value of the RNA_target attribute.
}
\description{
Getter for RNA_target attribute of miRNA_Gene class.
}
