% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data.R
\docType{data}
\name{rRNA}
\alias{rRNA}
\title{List of rRNAs}
\format{
A character vector with 4 elements.
}
\usage{
rRNA
}
\description{
This variable contains a list of possible rRNAs used for validation in the \code{rRNA_Gene} class.
}
\examples{
rRNA1 <- rRNA_Gene(id = 1, specific_RNA = "18s")

}
\keyword{datasets}
