% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_rRNA.R
\name{rRNA_ID,rRNA_Gene-method}
\alias{rRNA_ID,rRNA_Gene-method}
\title{Get rRNA_ID attribute}
\usage{
\S4method{rRNA_ID}{rRNA_Gene}(x)
}
\arguments{
\item{x}{An object of class rRNA_Gene.}
}
\value{
The value of the rRNA_ID attribute.
}
\description{
Getter for rRNA_ID attribute of rRNA_Gene class.
}
