% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_rRNA.R
\name{rRNA_ID<-,rRNA_Gene-method}
\alias{rRNA_ID<-,rRNA_Gene-method}
\title{Set rRNA_ID attribute}
\usage{
\S4method{rRNA_ID}{rRNA_Gene}(x) <- value
}
\arguments{
\item{x}{An object of class rRNA_Gene.}

\item{value}{The value to set for the rRNA_ID attribute.}
}
\value{
The modified object.
}
\description{
Setter for rRNA_ID attribute of rRNA_Gene class.
}
