% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_NON_Coding_gene_classes.R
\docType{class}
\name{sncRNA_Gene-class}
\alias{sncRNA_Gene-class}
\alias{sncRNA_Gene}
\title{sncRNA_Gene Class}
\value{
A \code{sncRNA_Gene} object.
}
\description{
A class to represent small non-coding RNA (sncRNA) genes.
}
\details{
Additional details:
The prototype are:
length_sncRNA = NA_real_, sncRNA_ID = NA_real_, types_RNA = "short".

The validity function checks the validity of the object by ensuring:
\itemize{
\item the length of the sncRNA is at most 200.
\item the type is correct (short).
}
}
\section{Slots}{

\describe{
\item{\code{length_sncRNA}}{A numeric value indicating the length of the sncRNA.}

\item{\code{sncRNA_ID}}{A numeric value representing the sncRNA ID.}

\item{\code{types_RNA}}{A character string representing the type of RNA.
Default is "short" and it's not possible to change it.}
}}

\section{Author}{

Elena Sasso \email{elena.sasso@mail.polimi.it}
}

\examples{
# Create a sncRNA_Gene object
sncRNA_gene <- sncRNA_Gene(id = 7, length_sncRNA = 80, sncRNA_ID = 1)

}
\seealso{
\code{\link{ncRNA_Gene}}
}
