% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_sncRNA.R
\name{sncRNA_ID<-,sncRNA_Gene-method}
\alias{sncRNA_ID<-,sncRNA_Gene-method}
\title{Set sncRNA_ID attribute}
\usage{
\S4method{sncRNA_ID}{sncRNA_Gene}(x) <- value
}
\arguments{
\item{x}{An object of class sncRNA_Gene.}

\item{value}{The value to set for the sncRNA_ID attribute.}
}
\value{
The modified object.
}
\description{
Setter for sncRNA_ID attribute of sncRNA_Gene class.
}
