% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_sncRNA.R
\name{sncRNA_ID}
\alias{sncRNA_ID}
\title{Get sncRNA_ID attribute}
\usage{
sncRNA_ID(x)
}
\arguments{
\item{x}{An object of class sncRNA_Gene.}
}
\value{
The value of the sncRNA_ID attribute.
}
\description{
Getter for sncRNA_ID attribute of sncRNA_Gene class.
}
