% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_NON_Coding_gene_classes.R
\docType{class}
\name{snoRNA_Gene-class}
\alias{snoRNA_Gene-class}
\alias{snoRNA_Gene}
\title{snoRNA_Gene Class}
\value{
A \code{snoRNA_Gene} object.
}
\description{
A class to represent small nucleolar RNA (snoRNA) genes.
}
\details{
Additional details:
The prototype are:
snoRNA_ID = NA_real_, gene_product = "snoRNA".

The validity function ensures the gene product is 'snoRNA'.
}
\section{Slots}{

\describe{
\item{\code{snoRNA_ID}}{A numeric value representing the snoRNA ID.}

\item{\code{gene_product}}{A character string representing the product of the gene.
Default is "snoRNA" and it's not possibile to change it.}
}}

\section{Author}{

Elena Sasso \email{elena.sasso@mail.polimi.it}
}

\examples{
# Create a snoRNA_Gene object
snoRNA_gene <- snoRNA_Gene(id = 9, snoRNA_ID = 1)

}
\seealso{
\code{\link{sncRNA_Gene}}
}
