% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_snoRNA_Gene.R
\name{snoRNA_ID<-}
\alias{snoRNA_ID<-}
\title{Set snoRNA_ID attribute}
\usage{
snoRNA_ID(x) <- value
}
\arguments{
\item{x}{An object of class snoRNA_Gene.}

\item{value}{The value to set for the snoRNA_ID attribute.}
}
\value{
The modified object.
}
\description{
Setter for snoRNA_ID attribute of snoRNA_Gene class.
}
