% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_snoRNA_Gene.R
\name{snoRNA_ID,snoRNA_Gene-method}
\alias{snoRNA_ID,snoRNA_Gene-method}
\title{Get snoRNA_ID attribute}
\usage{
\S4method{snoRNA_ID}{snoRNA_Gene}(x)
}
\arguments{
\item{x}{An object of class snoRNA_Gene.}
}
\value{
The value of the snoRNA_ID attribute.
}
\description{
Getter for snoRNA_ID attribute of snoRNA_Gene class.
}
