% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_NON_Coding_gene_classes.R
\docType{class}
\name{snRNA_Gene-class}
\alias{snRNA_Gene-class}
\alias{snRNA_Gene}
\title{snRNA_Gene Class}
\value{
A \code{snRNA_Gene} object.
}
\description{
A class to represent small nuclear RNA (snRNA) genes.
}
\details{
Additional details:
The prototype are:
snRNA_ID = NA_real_, gene_product = "snRNA".

The validity function ensures the gene product is 'snRNA'.
}
\section{Slots}{

\describe{
\item{\code{snRNA_ID}}{A numeric value representing the snRNA ID.}

\item{\code{gene_product}}{A character string representing the product of the gene.
Default is "snRNA" and it's not possibile to change it.}
}}

\section{Author}{

Elena Sasso \email{elena.sasso@mail.polimi.it}
}

\examples{
# Create a snRNA_Gene object
snRNA_gene <- snRNA_Gene(id = 10, snRNA_ID = 1)

}
\seealso{
\code{\link{sncRNA_Gene}}
}
