% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_snRNA_Gene.R
\name{snRNA_ID<-}
\alias{snRNA_ID<-}
\title{Set snRNA_ID attribute}
\usage{
snRNA_ID(x) <- value
}
\arguments{
\item{x}{An object of class snRNA_Gene.}

\item{value}{The value to set for the snRNA_ID attribute.}
}
\value{
The modified object.
}
\description{
Setter for snRNA_ID attribute of snRNA_Gene class.
}
