% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_snRNA_Gene.R
\name{snRNA_ID,snRNA_Gene-method}
\alias{snRNA_ID,snRNA_Gene-method}
\title{Get snRNA_ID attribute}
\usage{
\S4method{snRNA_ID}{snRNA_Gene}(x)
}
\arguments{
\item{x}{An object of class snRNA_Gene.}
}
\value{
The value of the snRNA_ID attribute.
}
\description{
Getter for snRNA_ID attribute of snRNA_Gene class.
}
