% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_tRNA.R
\name{specific_aa<-,tRNA_Gene-method}
\alias{specific_aa<-,tRNA_Gene-method}
\title{Set specific_aa attribute}
\usage{
\S4method{specific_aa}{tRNA_Gene}(x) <- value
}
\arguments{
\item{x}{An object of class tRNA_Gene.}

\item{value}{The value to set for the specific_aa attribute.}
}
\value{
The modified object.
}
\description{
Setter for specific_aa attribute of tRNA_Gene class.
}
