% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_tRNA.R
\name{specific_aa,tRNA_Gene-method}
\alias{specific_aa,tRNA_Gene-method}
\title{Get specific_aa attribute}
\usage{
\S4method{specific_aa}{tRNA_Gene}(x)
}
\arguments{
\item{x}{An object of class tRNA_Gene.}
}
\value{
The value of the specific_aa attribute.
}
\description{
Getter for specific_aa attribute of tRNA_Gene class.
}
