% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_rRNA.R
\name{specific_RNA,rRNA_Gene-method}
\alias{specific_RNA,rRNA_Gene-method}
\title{Get specific_RNA attribute}
\usage{
\S4method{specific_RNA}{rRNA_Gene}(x)
}
\arguments{
\item{x}{An object of class rRNA_Gene.}
}
\value{
The value of the specific_RNA attribute.
}
\description{
Getter for specific_RNA attribute of rRNA_Gene class.
}
