% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_rRNA.R
\name{specific_RNA<-}
\alias{specific_RNA<-}
\title{Set specific_RNA attribute}
\usage{
specific_RNA(x) <- value
}
\arguments{
\item{x}{An object of class rRNA_Gene.}

\item{value}{The value to set for the specific_RNA attribute.}
}
\value{
The modified object.
}
\description{
Setter for specific_RNA attribute of rRNA_Gene class.
}
