% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_Gene.R
\name{start<-,Gene-method}
\alias{start<-,Gene-method}
\title{Set start attribute}
\usage{
\S4method{start}{Gene}(x) <- value
}
\arguments{
\item{x}{An object of class Gene.}

\item{value}{The value to set for the start attribute.}
}
\value{
The modified object.
}
\description{
Setter for start attribute of Gene class.
}
