% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{start<-}
\alias{start<-}
\title{Set start attribute}
\usage{
start(x) <- value
}
\arguments{
\item{x}{An object of class Gene.}

\item{value}{The value to set for the start attribute.}
}
\value{
The modified object.
}
\description{
Setter for start attribute of Gene class.
}
