% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_Gene.R
\name{start}
\alias{start}
\title{Get start attribute}
\usage{
start(x, ...)
}
\arguments{
\item{x}{An object of class Gene.}
}
\value{
The value of the start attribute.
}
\description{
Getter for start attribute of Gene class.
}
