% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_Gene.R
\name{strand,Gene-method}
\alias{strand,Gene-method}
\title{Get strand attribute}
\usage{
\S4method{strand}{Gene}(x)
}
\arguments{
\item{x}{An object of class Gene.}
}
\value{
The value of the strand attribute.
}
\description{
Getter for strand attribute of Gene class.
}
