% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{strand<-}
\alias{strand<-}
\title{Set strand attribute}
\usage{
strand(x) <- value
}
\arguments{
\item{x}{An object of class Gene.}

\item{value}{The value to set for the strand attribute.}
}
\value{
The modified object.
}
\description{
Setter for strand attribute of Gene class.
}
