% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{strand}
\alias{strand}
\title{Get strand attribute}
\usage{
strand(x)
}
\arguments{
\item{x}{An object of class Gene.}
}
\value{
The value of the strand attribute.
}
\description{
Getter for strand attribute of Gene class.
}
