% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_tRNA.R
\name{tRNA_ID<-,tRNA_Gene-method}
\alias{tRNA_ID<-,tRNA_Gene-method}
\title{Set tRNA_ID attribute}
\usage{
\S4method{tRNA_ID}{tRNA_Gene}(x) <- value
}
\arguments{
\item{x}{An object of class tRNA_Gene.}

\item{value}{The value to set for the tRNA_ID attribute.}
}
\value{
The modified object.
}
\description{
Setter for tRNA_ID attribute of tRNA_Gene class.
}
