% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_tRNA.R
\name{tRNA_ID,tRNA_Gene-method}
\alias{tRNA_ID,tRNA_Gene-method}
\title{Get tRNA_ID attribute}
\usage{
\S4method{tRNA_ID}{tRNA_Gene}(x)
}
\arguments{
\item{x}{An object of class tRNA_Gene.}
}
\value{
The value of the tRNA_ID attribute.
}
\description{
Getter for tRNA_ID attribute of tRNA_Gene class.
}
