% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_Gene.R
\name{TSS,Gene-method}
\alias{TSS,Gene-method}
\title{Get TSS attribute}
\usage{
\S4method{TSS}{Gene}(x)
}
\arguments{
\item{x}{An object of class Gene.}
}
\value{
The value of the TSS attribute.
}
\description{
Getter for TSS attribute of Gene class.
}
