% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_Gene.R
\name{TSS<-}
\alias{TSS<-}
\title{Set TSS attribute}
\usage{
TSS(x) <- value
}
\arguments{
\item{x}{An object of class Gene.}

\item{value}{The value to set for the TSS attribute.}
}
\value{
The modified object.
}
\description{
Setter for TSS attribute of Gene class.
}
