% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{TSS}
\alias{TSS}
\title{Get TSS attribute}
\usage{
TSS(x)
}
\arguments{
\item{x}{An object of class Gene.}
}
\value{
The value of the TSS attribute.
}
\description{
Getter for TSS attribute of Gene class.
}
