% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_Gene.R
\name{TTS<-,Gene-method}
\alias{TTS<-,Gene-method}
\title{Set TTS attribute}
\usage{
\S4method{TTS}{Gene}(x) <- value
}
\arguments{
\item{x}{An object of class Gene.}

\item{value}{The value to set for the TTS attribute.}
}
\value{
The modified object.
}
\description{
Setter for TTS attribute of Gene class.
}
