% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{TTS}
\alias{TTS}
\title{Get TTS attribute}
\usage{
TTS(x)
}
\arguments{
\item{x}{An object of class Gene.}
}
\value{
The value of the TTS attribute.
}
\description{
Getter for TTS attribute of Gene class.
}
