% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_lncRNA.R
\name{type_RNA<-}
\alias{type_RNA<-}
\title{Set type_RNA attribute}
\usage{
type_RNA(x) <- value
}
\arguments{
\item{x}{An object of class lncRNA_Gene.}

\item{value}{The value to set for the type_RNA attribute.}
}
\value{
The modified object.
}
\description{
Setter for type_RNA attribute of lncRNA_Gene class.
}
