% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_lncRNA.R
\name{type_RNA}
\alias{type_RNA}
\title{Get type_RNA attribute}
\usage{
type_RNA(x)
}
\arguments{
\item{x}{An object of class lncRNA_Gene.}
}
\value{
The value of the type_RNA attribute.
}
\description{
Getter for type_RNA attribute of lncRNA_Gene class.
}
