% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_sncRNA.R
\name{types_RNA<-,sncRNA_Gene-method}
\alias{types_RNA<-,sncRNA_Gene-method}
\title{Set types_RNA attribute}
\usage{
\S4method{types_RNA}{sncRNA_Gene}(x) <- value
}
\arguments{
\item{x}{An object of class sncRNA_Gene.}

\item{value}{The value to set for the types_RNA attribute.}
}
\value{
The modified object.
}
\description{
Setter for types_RNA attribute of sncRNA_Gene class.
}
