% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters_sncRNA.R
\name{types_RNA,sncRNA_Gene-method}
\alias{types_RNA,sncRNA_Gene-method}
\title{Get types_RNA attribute}
\usage{
\S4method{types_RNA}{sncRNA_Gene}(x)
}
\arguments{
\item{x}{An object of class sncRNA_Gene.}
}
\value{
The value of the types_RNA attribute.
}
\description{
Getter for types_RNA attribute of sncRNA_Gene class.
}
