% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_getter_setter.R
\name{generateGetterSetterCode_for_small}
\alias{generateGetterSetterCode_for_small}
\title{Generate Getter and Setter Code for small_RNAs genes classes
This function generates getter and setter methods for all attributes of the
small_RNAs genes classes, writing the code to a specified file.}
\usage{
generateGetterSetterCode_for_small(class_name, file_path)
}
\arguments{
\item{class_name}{The name of the S4 class for which the getter and setter
methods are to be generated.}

\item{file_path}{The file path where the generated code will be written.}
}
\value{
This function does not return a value. It writes the generated
code directly to the specified file.
}
\description{
Generate Getter and Setter Code for small_RNAs genes classes
This function generates getter and setter methods for all attributes of the
small_RNAs genes classes, writing the code to a specified file.
}
\details{
This function generates Roxygen2 documentation for each getter
and setter method, making them ready for inclusion in an R package.
}
