% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_Functions.R
\name{lengthProduct}
\alias{lengthProduct}
\title{LengthProduct Generic Function}
\usage{
lengthProduct(classes)
}
\arguments{
\item{classes}{classes that contains a length attribute for the gene
product.}
}
\value{
The length of the gene product.
}
\description{
A generic function to calculate the length of the product of a gene.
}
